requirements.txt
py-consul
psycopg
psycopg-binary
redis
pika
